Rename csv_utils to _csv_utils and fix imports
